Remove dependency on outline-ss-server and cleanup
