+++ outline/connectivity.go
@@ -1,4 +1,18 @@
-package shadowsocks
+// Copyright 2023 The Outline Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package outline
 
 import (
 	"context"
@@ -7,7 +21,7 @@ import (
 	"net/http"
 	"time"
 
-	onet "github.com/Jigsaw-Code/outline-ss-server/net"
+	"github.com/Jigsaw-Code/outline-internal-sdk/transport"
 )
 
 // TODO: make these values configurable by exposing a struct with the connectivity methods.
@@ -28,10 +42,10 @@ type ReachabilityError struct {
 	error
 }
 
-// CheckUDPConnectivityWithDNS determines whether the Shadowsocks proxy represented by `client` and
+// checkUDPConnectivityWithDNS determines whether the Shadowsocks proxy represented by `client` and
 // the network support UDP traffic by issuing a DNS query though a resolver at `resolverAddr`.
 // Returns nil on success or an error on failure.
-func CheckUDPConnectivityWithDNS(client onet.PacketListener, resolverAddr net.Addr) error {
+func CheckUDPConnectivityWithDNS(client transport.PacketListener, resolverAddr net.Addr) error {
 	conn, err := client.ListenPacket(context.Background())
 	if err != nil {
 		return err
@@ -56,11 +70,11 @@ func CheckUDPConnectivityWithDNS(client onet.PacketListener, resolverAddr net.Ad
 	return errors.New("UDP connectivity check timed out")
 }
 
-// CheckTCPConnectivityWithHTTP determines whether the proxy is reachable over TCP and validates the
+// checkTCPConnectivityWithHTTP determines whether the proxy is reachable over TCP and validates the
 // client's authentication credentials by performing an HTTP HEAD request to `targetURL`, which must
 // be of the form: http://[host](:[port])(/[path]). Returns nil on success, error if `targetURL` is
 // invalid, AuthenticationError or ReachabilityError on connectivity failure.
-func CheckTCPConnectivityWithHTTP(dialer onet.StreamDialer, targetURL string) error {
+func CheckTCPConnectivityWithHTTP(dialer transport.StreamDialer, targetURL string) error {
 	deadline := time.Now().Add(tcpTimeout)
 	ctx, cancel := context.WithDeadline(context.Background(), deadline)
 	defer cancel()
