+++ outline/connectivity_test.go
@@ -1,4 +1,18 @@
-package shadowsocks
+// Copyright 2023 The Outline Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package outline
 
 import (
 	"context"
@@ -8,8 +22,8 @@ import (
 	"testing"
 	"time"
 
-	onet "github.com/Jigsaw-Code/outline-ss-server/net"
-	ss "github.com/Jigsaw-Code/outline-ss-server/shadowsocks"
+	"github.com/Jigsaw-Code/outline-internal-sdk/transport"
+	"github.com/Jigsaw-Code/outline-internal-sdk/transport/shadowsocks"
 )
 
 func TestCheckUDPConnectivityWithDNS_Success(t *testing.T) {
@@ -65,7 +79,7 @@ type fakeSSClient struct {
 	failUDP            bool
 }
 
-func (c *fakeSSClient) Dial(_ context.Context, raddr string) (onet.DuplexConn, error) {
+func (c *fakeSSClient) Dial(_ context.Context, raddr string) (transport.StreamConn, error) {
 	if c.failReachability {
 		return nil, &net.OpError{}
 	}
@@ -80,7 +94,7 @@ func (c *fakeSSClient) ListenPacket(_ context.Context) (net.PacketConn, error) {
 	failRead := c.failAuthentication || c.failUDP || c.failReachability
 	return &fakePacketConn{PacketConn: conn, failRead: failRead}, nil
 }
-func (c *fakeSSClient) SetTCPSaltGenerator(salter ss.SaltGenerator) {
+func (c *fakeSSClient) SetTCPSaltGenerator(salter shadowsocks.SaltGenerator) {
 }
 
 // Fake PacketConn that fails `ReadFrom` calls when `failRead` is true.
@@ -104,7 +118,7 @@ func (c *fakePacketConn) ReadFrom(b []byte) (int, net.Addr, error) {
 
 // Fake DuplexConn that fails `Read` calls when `failRead` is true.
 type fakeDuplexConn struct {
-	onet.DuplexConn
+	transport.StreamConn
 	failRead bool
 }
