+++ outline/electron/main.go
@@ -24,8 +24,8 @@ import (
 	"syscall"
 	"time"
 
-	oss "github.com/Jigsaw-Code/outline-go-tun2socks/outline/shadowsocks"
-	"github.com/Jigsaw-Code/outline-go-tun2socks/shadowsocks"
+	"github.com/Jigsaw-Code/outline-go-tun2socks/outline"
+	"github.com/Jigsaw-Code/outline-go-tun2socks/outline/shadowsocks"
 	"github.com/eycorsican/go-tun2socks/common/log"
 	_ "github.com/eycorsican/go-tun2socks/common/log/simple" // Register a simple logger.
 	"github.com/eycorsican/go-tun2socks/core"
@@ -86,19 +86,19 @@ func main() {
 	// Validate proxy flags
 	if *args.proxyHost == "" {
 		log.Errorf("Must provide a Shadowsocks proxy host name or IP address")
-		os.Exit(oss.IllegalConfiguration)
+		os.Exit(shadowsocks.IllegalConfiguration)
 	} else if *args.proxyPort <= 0 || *args.proxyPort > 65535 {
 		log.Errorf("Must provide a valid Shadowsocks proxy port [1:65535]")
-		os.Exit(oss.IllegalConfiguration)
+		os.Exit(shadowsocks.IllegalConfiguration)
 	} else if *args.proxyPassword == "" {
 		log.Errorf("Must provide a Shadowsocks proxy password")
-		os.Exit(oss.IllegalConfiguration)
+		os.Exit(shadowsocks.IllegalConfiguration)
 	} else if *args.proxyCipher == "" {
 		log.Errorf("Must provide a Shadowsocks proxy encryption cipher")
-		os.Exit(oss.IllegalConfiguration)
+		os.Exit(shadowsocks.IllegalConfiguration)
 	}
 
-	config := oss.Config{
+	config := shadowsocks.Config{
 		Host:       *args.proxyHost,
 		Port:       *args.proxyPort,
 		Password:   *args.proxyPassword,
@@ -112,19 +112,19 @@ func main() {
 	for i, r := range prefixRunes {
 		if (r & 0xFF) != r {
 			log.Errorf("Character out of range: %r", r)
-			os.Exit(oss.IllegalConfiguration)
+			os.Exit(shadowsocks.IllegalConfiguration)
 		}
 		config.Prefix[i] = byte(r)
 	}
 
-	client, err := oss.NewClient(&config)
+	client, err := shadowsocks.NewClient(&config)
 	if err != nil {
 		log.Errorf("Failed to construct Shadowsocks client: %v", err)
-		os.Exit(oss.IllegalConfiguration)
+		os.Exit(shadowsocks.IllegalConfiguration)
 	}
 
 	if *args.checkConnectivity {
-		connErrCode, err := oss.CheckConnectivity(client)
+		connErrCode, err := shadowsocks.CheckConnectivity(client)
 		log.Debugf("Connectivity checks error code: %v", connErrCode)
 		if err != nil {
 			log.Errorf("Failed to perform connectivity checks: %v", err)
@@ -137,19 +137,19 @@ func main() {
 	tunDevice, err := tun.OpenTunDevice(*args.tunName, *args.tunAddr, *args.tunGw, *args.tunMask, dnsResolvers, persistTun)
 	if err != nil {
 		log.Errorf("Failed to open TUN device: %v", err)
-		os.Exit(oss.SystemMisconfigured)
+		os.Exit(shadowsocks.SystemMisconfigured)
 	}
 	// Output packets to TUN device
 	core.RegisterOutputFn(tunDevice.Write)
 
 	// Register TCP and UDP connection handlers
-	core.RegisterTCPConnHandler(shadowsocks.NewTCPHandler(client))
+	core.RegisterTCPConnHandler(outline.NewTCPHandler(client))
 	if *args.dnsFallback {
 		// UDP connectivity not supported, fall back to DNS over TCP.
 		log.Debugf("Registering DNS fallback UDP handler")
 		core.RegisterUDPConnHandler(dnsfallback.NewUDPHandler())
 	} else {
-		core.RegisterUDPConnHandler(shadowsocks.NewUDPHandler(client, udpTimeout))
+		core.RegisterUDPConnHandler(outline.NewUDPHandler(client, udpTimeout))
 	}
 
 	// Configure LWIP stack to receive input data from the TUN device
@@ -158,7 +158,7 @@ func main() {
 		_, err := io.CopyBuffer(lwipWriter, tunDevice, make([]byte, mtu))
 		if err != nil {
 			log.Errorf("Failed to write data to network stack: %v", err)
-			os.Exit(oss.Unexpected)
+			os.Exit(shadowsocks.Unexpected)
 		}
 	}()
