+++ outline/tunnel.go
@@ -23,10 +23,9 @@ import (
 	"github.com/eycorsican/go-tun2socks/core"
 	"github.com/eycorsican/go-tun2socks/proxy/dnsfallback"
 
-	oss "github.com/Jigsaw-Code/outline-go-tun2socks/shadowsocks"
+	"github.com/Jigsaw-Code/outline-internal-sdk/transport"
 
 	"github.com/Jigsaw-Code/outline-go-tun2socks/tunnel"
-	onet "github.com/Jigsaw-Code/outline-ss-server/net"
 )
 
 // Tunnel represents a tunnel from a TUN device to a server.
@@ -42,8 +41,8 @@ type Tunnel interface {
 type outlinetunnel struct {
 	tunnel.Tunnel
 	lwipStack    core.LWIPStack
-	streamDialer onet.StreamDialer
-	packetDialer onet.PacketListener
+	streamDialer transport.StreamDialer
+	packetDialer transport.PacketListener
 	isUDPEnabled bool // Whether the tunnel supports proxying UDP.
 }
 
@@ -55,7 +54,7 @@ type outlinetunnel struct {
 // `cipher` is the encryption cipher used by the Shadowsocks proxy.
 // `isUDPEnabled` indicates if the Shadowsocks proxy and the network support proxying UDP traffic.
 // `tunWriter` is used to output packets back to the TUN device.  OutlineTunnel.Disconnect() will close `tunWriter`.
-func NewTunnel(streamDialer onet.StreamDialer, packetDialer onet.PacketListener, isUDPEnabled bool, tunWriter io.WriteCloser) (Tunnel, error) {
+func NewTunnel(streamDialer transport.StreamDialer, packetDialer transport.PacketListener, isUDPEnabled bool, tunWriter io.WriteCloser) (Tunnel, error) {
 	if tunWriter == nil {
 		return nil, errors.New("Must provide a TUN writer")
 	}
@@ -71,7 +70,7 @@ func NewTunnel(streamDialer onet.StreamDialer, packetDialer onet.PacketListener,
 
 func (t *outlinetunnel) UpdateUDPSupport() bool {
 	resolverAddr := &net.UDPAddr{IP: net.ParseIP("1.1.1.1"), Port: 53}
-	isUDPEnabled := oss.CheckUDPConnectivityWithDNS(t.packetDialer, resolverAddr) == nil
+	isUDPEnabled := CheckUDPConnectivityWithDNS(t.packetDialer, resolverAddr) == nil
 	if t.isUDPEnabled != isUDPEnabled {
 		t.isUDPEnabled = isUDPEnabled
 		t.lwipStack.Close() // Close existing connections to avoid using the previous handlers.
@@ -85,10 +84,10 @@ func (t *outlinetunnel) UpdateUDPSupport() bool {
 func (t *outlinetunnel) registerConnectionHandlers() {
 	var udpHandler core.UDPConnHandler
 	if t.isUDPEnabled {
-		udpHandler = oss.NewUDPHandler(t.packetDialer, 30*time.Second)
+		udpHandler = NewUDPHandler(t.packetDialer, 30*time.Second)
 	} else {
 		udpHandler = dnsfallback.NewUDPHandler()
 	}
-	core.RegisterTCPConnHandler(oss.NewTCPHandler(t.streamDialer))
+	core.RegisterTCPConnHandler(NewTCPHandler(t.streamDialer))
 	core.RegisterUDPConnHandler(udpHandler)
 }
