+++ outline/udp.go
@@ -1,4 +1,18 @@
-package shadowsocks
+// Copyright 2023 The Outline Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package outline
 
 import (
 	"context"
@@ -7,7 +21,7 @@ import (
 	"sync"
 	"time"
 
-	onet "github.com/Jigsaw-Code/outline-ss-server/net"
+	transport "github.com/Jigsaw-Code/outline-internal-sdk/transport"
 	"github.com/eycorsican/go-tun2socks/core"
 )
 
@@ -16,7 +30,7 @@ type udpHandler struct {
 	sync.Mutex
 
 	// Used to establish connections to the proxy
-	listener onet.PacketListener
+	listener transport.PacketListener
 
 	// How long to wait for a packet from the proxy. Longer than this and the connection
 	// is closed.
@@ -30,7 +44,7 @@ type udpHandler struct {
 //
 // `client` provides the Shadowsocks functionality.
 // `timeout` is the UDP read and write timeout.
-func NewUDPHandler(dialer onet.PacketListener, timeout time.Duration) core.UDPConnHandler {
+func NewUDPHandler(dialer transport.PacketListener, timeout time.Duration) core.UDPConnHandler {
 	return &udpHandler{
 		listener: dialer,
 		timeout:  timeout,
